fix duplicate symbol error
